+++ Projects/KidsCare_UI/src/main/java/kr/ac/kopo/kidscare/model/KCUserPost.java
@@ -6,7 +6,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 
 public class KCUserPost {
-	private Long userPostId;
+	private Integer userPostId;
 	private String username;
 	@DateTimeFormat(pattern="yyyyMMddHHmmss")
 	private String postdate;
@@ -22,10 +22,12 @@ public class KCUserPost {
 		this.contents = contents;
 	}
 	
+
 	public Long getUserPostId() {
+
 		return userPostId;
 	}
-	public void setUserPostId(Long userPostId) {
+	public void setUserPostId(Integer userPostId) {
 		this.userPostId = userPostId;
 	}
 	public String getUsername() {
